fix: move mocks to their respective packages
