+++ internal/adapter/email/sendgrid_test.go
@@ -3,10 +3,8 @@ package email_test
 import (
 	"github.com/bxcodec/faker/v4"
 	"github.com/mohsenabedy91/polyglot-sentences/internal/adapter/email"
-	emailmocks "github.com/mohsenabedy91/polyglot-sentences/internal/adapter/email/mocks"
 	"github.com/mohsenabedy91/polyglot-sentences/internal/core/config"
 	"github.com/mohsenabedy91/polyglot-sentences/pkg/logger"
-	logmocks "github.com/mohsenabedy91/polyglot-sentences/pkg/logger/mocks"
 	"github.com/mohsenabedy91/polyglot-sentences/pkg/serviceerror"
 	"github.com/sendgrid/rest"
 	"github.com/stretchr/testify/mock"
@@ -15,8 +13,8 @@ import (
 )
 
 func TestSend_Successful(t *testing.T) {
-	mockLogger := new(logmocks.MockLogger)
-	mockClient := new(emailmocks.MockClient)
+	mockLogger := new(logger.MockLogger)
+	mockClient := new(email.MockClient)
 
 	conf := config.SendGrid{
 		Name:    "Test",
@@ -48,8 +46,8 @@ func TestSend_Successful(t *testing.T) {
 }
 
 func TestSend_Failure(t *testing.T) {
-	mockLogger := new(logmocks.MockLogger)
-	mockClient := new(emailmocks.MockClient)
+	mockLogger := new(logger.MockLogger)
+	mockClient := new(email.MockClient)
 
 	conf := config.SendGrid{
 		Name:    "Test",
@@ -80,8 +78,8 @@ func TestSend_Failure(t *testing.T) {
 }
 
 func TestSend_LoggingSuccess(t *testing.T) {
-	mockLogger := new(logmocks.MockLogger)
-	mockClient := new(emailmocks.MockClient)
+	mockLogger := new(logger.MockLogger)
+	mockClient := new(email.MockClient)
 
 	conf := config.SendGrid{
 		Name:    "Test",
@@ -122,8 +120,8 @@ func TestSend_LoggingSuccess(t *testing.T) {
 }
 
 func TestSend_LoggingFailure(t *testing.T) {
-	mockLogger := new(logmocks.MockLogger)
-	mockClient := new(emailmocks.MockClient)
+	mockLogger := new(logger.MockLogger)
+	mockClient := new(email.MockClient)
 
 	conf := config.SendGrid{
 		Name:    "Test",
