+++ pkg/logger/mock_logger.go
@@ -1,7 +1,6 @@
-package mocks
+package logger
 
 import (
-	"github.com/mohsenabedy91/polyglot-sentences/pkg/logger"
 	"github.com/stretchr/testify/mock"
 )
 
@@ -12,7 +11,7 @@ type MockLogger struct {
 func (r *MockLogger) Init(appName string) {
 }
 
-func (r *MockLogger) Debug(category logger.Category, subCategory logger.SubCategory, message string, extra map[logger.ExtraKey]interface{}) {
+func (r *MockLogger) Debug(category Category, subCategory SubCategory, message string, extra map[ExtraKey]interface{}) {
 	r.Called(category, subCategory, message, extra)
 }
 
@@ -20,7 +19,7 @@ func (r *MockLogger) DebugF(template string, args ...interface{}) {
 	r.Called(template, args)
 }
 
-func (r *MockLogger) Info(category logger.Category, subCategory logger.SubCategory, message string, extra map[logger.ExtraKey]interface{}) {
+func (r *MockLogger) Info(category Category, subCategory SubCategory, message string, extra map[ExtraKey]interface{}) {
 	r.Called(category, subCategory, message, extra)
 }
 
@@ -28,7 +27,7 @@ func (r *MockLogger) InfoF(template string, args ...interface{}) {
 	r.Called(template, args)
 }
 
-func (r *MockLogger) Warn(category logger.Category, subCategory logger.SubCategory, message string, extra map[logger.ExtraKey]interface{}) {
+func (r *MockLogger) Warn(category Category, subCategory SubCategory, message string, extra map[ExtraKey]interface{}) {
 	r.Called(category, subCategory, message, extra)
 }
 
@@ -36,7 +35,7 @@ func (r *MockLogger) WarnF(template string, args ...interface{}) {
 	r.Called(template, args)
 }
 
-func (r *MockLogger) Error(category logger.Category, subCategory logger.SubCategory, message string, extra map[logger.ExtraKey]interface{}) {
+func (r *MockLogger) Error(category Category, subCategory SubCategory, message string, extra map[ExtraKey]interface{}) {
 	r.Called(category, subCategory, message, extra)
 }
 
@@ -44,7 +43,7 @@ func (r *MockLogger) ErrorF(template string, args ...interface{}) {
 	r.Called(template, args)
 }
 
-func (r *MockLogger) Fatal(category logger.Category, subCategory logger.SubCategory, message string, extra map[logger.ExtraKey]interface{}) {
+func (r *MockLogger) Fatal(category Category, subCategory SubCategory, message string, extra map[ExtraKey]interface{}) {
 	r.Called(category, subCategory, message, extra)
 }
