+++ pkg/oauth/google_test.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 	"errors"
 	"github.com/mohsenabedy91/polyglot-sentences/internal/core/config"
-	mocklogger "github.com/mohsenabedy91/polyglot-sentences/pkg/logger/mocks"
+	"github.com/mohsenabedy91/polyglot-sentences/pkg/logger"
 	"github.com/mohsenabedy91/polyglot-sentences/pkg/oauth"
 	"github.com/mohsenabedy91/polyglot-sentences/pkg/serviceerror"
 	"github.com/stretchr/testify/mock"
@@ -119,7 +119,7 @@ func TestUserGoogleInfo(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
-			mockLogger := new(mocklogger.MockLogger)
+			mockLogger := new(logger.MockLogger)
 			mockLogger.On("Error", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil)
 
 			mockClientProvider := new(MockHTTPClientProvider)
